Move `createLocalSnippetStoreV2` to `wa-sqlite.ts`
