+++ src/lib/components/card-v2/store.ts
@@ -1,8 +1,8 @@
 import { derived } from 'svelte/store'
 import { globalStateStore } from '$lib/utitlities/global'
-import { createLocalSnippetStoreV2 } from '$lib/utitlities/persistence'
 import { localDb } from '$lib/sqlite/global'
 import { migrationStateStore } from '$lib/sqlite/migration'
+import { createLocalSnippetStoreV2 } from '$lib/sqlite/wa-sqlite'
 
 export const localSnippetsStore = await createLocalSnippetStoreV2(migrationStateStore, globalStateStore, localDb)
 export const displaySnippetsStore = derived(
