+++ src/lib/sqlite/wa-sqlite.ts
@@ -7,9 +7,26 @@ import { MemoryAsyncVFS } from 'wa-sqlite/src/examples/MemoryAsyncVFS.js'
 import { MemoryVFS } from 'wa-sqlite/src/examples/MemoryVFS.js'
 // @ts-ignore
 import { OriginPrivateFileSystemVFS } from 'wa-sqlite/src/examples/OriginPrivateFileSystemVFS.js'
-import { drizzle } from 'drizzle-orm/sqlite-proxy';
-import type { MigrationQueryMap, QueriesStringMap } from '$lib/sqlite/migration';
-import { sql } from 'drizzle-orm';
+import { drizzle } from 'drizzle-orm/sqlite-proxy'
+import type { MigrationQueryMap, QueriesStringMap } from '$lib/sqlite/migration'
+import type { MigrationState } from '$lib/sqlite/migration'
+import { derived, writable } from 'svelte/store'
+import type { Invalidator, Subscriber, Unsubscriber, Writable } from 'svelte/store'
+import type { SqliteRemoteDatabase } from 'drizzle-orm/sqlite-proxy'
+import {
+  clearTags,
+  deleteSnippet as deleteSnippet_,
+  querySnippetsByFolderId,
+  queryTagsBySnippetIds,
+  upsertSnippet,
+  upsertTags
+} from '$lib/sqlite/queries';
+import {
+  buildTagsMap,
+  dbSnippetToDisplaySnippet,
+  displaySnippetToDbSnippet
+} from '$lib/utitlities/data-transformation';
+import type { GlobalState, Snippet, SnippetStore } from '$lib/utitlities/persistence';
 
 export type WASqliteExecutor = {
   execute(query: string, ...params: SQLiteCompatibleType[]): Promise<SQLiteCompatibleType[][]>
@@ -136,5 +153,104 @@ export async function migrateLocal(
     }
     currentUserVersion += 1
     await executor.execute(`PRAGMA user_version = ${currentUserVersion}`)
+  }
+}
+
+export async function createLocalSnippetStoreV2(
+  migrationStateStore: Writable<MigrationState>,
+  globalStateStore: Writable<GlobalState>,
+  db: SqliteRemoteDatabase,
+): Promise<SnippetStore> {
+  let snippets: Snippet[] = []
+  let folderId = 'default'
+  const store = writable(snippets)
+  const stores = derived(
+    [globalStateStore, migrationStateStore],
+    ([globalState, migrationState]: [GlobalState, MigrationState]) => {
+      folderId = globalState.folderId
+      return [globalState, migrationState]
+    }
+  )
+  const pairUnsubscribeFn = stores.subscribe(
+    // TODO: fix the typing of `globalState` and `migrationState`, which both
+    //       have the type `GlobalState | MigrationState`
+    async ([globalState, migrationState]) => {
+      if (migrationState === 'done') {
+        const dbSnippets = await querySnippetsByFolderId(db, (globalState as GlobalState).folderId)
+        const snippetIds = dbSnippets.map(snippet => snippet.id)
+        const tags = await queryTagsBySnippetIds(db, snippetIds)
+        const tagsMap = buildTagsMap(tags)
+        snippets = dbSnippets.map(
+          (dbSnippet) => dbSnippetToDisplaySnippet(dbSnippet, tagsMap)
+        )
+        store.set(snippets)
+      }
+    }
+  )
+
+  return {
+    subscribe(
+      run: Subscriber<Snippet[]>,
+      invalidate?: Invalidator<Snippet[]>,
+    ): Unsubscriber {
+      const baseUnsubscribeFn = store.subscribe(run, invalidate)
+
+      return () => {
+        pairUnsubscribeFn()
+        baseUnsubscribeFn()
+      }
+    },
+    clone: async (snippet: Snippet) => {
+      const clonedSnippet: Snippet = {
+        ...snippet,
+        id: crypto.randomUUID(),
+        position: snippets.length + 1,
+        tags: snippet.tags.slice(),
+        createdAt: new Date().getTime(),
+        updatedAt: new Date().getTime(),
+      }
+      const dbSnippet = displaySnippetToDbSnippet(folderId, clonedSnippet)
+      await upsertSnippet(db, dbSnippet)
+      if (snippet.tags.length > 0) {
+        await upsertTags(db, clonedSnippet.id, snippet.tags)
+      }
+      snippets.push(clonedSnippet)
+
+      store.set(snippets)
+    },
+    upsert: async (snippet: Snippet) => {
+      const dbSnippet = displaySnippetToDbSnippet(folderId, snippet)
+      await upsertSnippet(db, dbSnippet)
+      if (snippet.tags && snippet.tags.length > 0) {
+        await clearTags(db, snippet.id)
+        await upsertTags(db, snippet.id, snippet.tags)
+      }
+      const index = snippets.findIndex(snippet_ => snippet_.id === snippet.id)
+      if (index === -1) {
+        snippets.push(snippet)
+        store.set(snippets)
+        return
+      }
+
+      snippets[index] = snippet
+      store.set(snippets)
+    },
+    remove: async (id: string) => {
+      await deleteSnippet_(db, id)
+      const index = snippets.findIndex(snippet => snippet.id === id)
+      snippets.splice(index, 1)
+
+      store.set(snippets)
+    },
+    move: async (movingSnippet: Snippet, sourceFolderId: string, destinationFolderId: string) => {
+      const dbSnippet = displaySnippetToDbSnippet(sourceFolderId, movingSnippet)
+      dbSnippet.folderId = destinationFolderId
+      await upsertSnippet(db, dbSnippet)
+
+      const index = snippets.findIndex(snippet => snippet.id === movingSnippet.id)
+      snippets.splice(index, 1)
+
+      store.set(snippets)
+    },
   }
 }
