+++ src/lib/utitlities/persistence.ts
@@ -1,23 +1,8 @@
-import { derived, writable } from 'svelte/store'
-import type { Readable, Writable, Invalidator, Subscriber, Unsubscriber } from 'svelte/store'
+import type { Readable } from 'svelte/store'
 import { aesGcmDecrypt, aesGcmEncrypt } from '$lib/utitlities/encryption'
 import type { SqliteRemoteDatabase } from 'drizzle-orm/sqlite-proxy'
-import type { MigrationState } from '$lib/sqlite/migration'
-import {
-  querySnippetsByFolderId,
-  queryTagsBySnippetIds,
-  upsertSnippet,
-  deleteSnippet as deleteSnippet_,
-  upsertTags,
-  clearTags,
-} from '$lib/sqlite/queries';
-import {
-  buildTagsMap,
-  dbSnippetToDisplaySnippet,
-  displaySnippetToDbSnippet
-} from '$lib/utitlities/data-transformation';
-import { folders, snippet_tags, snippets } from '$lib/sqlite/schema';
-import { sql } from 'drizzle-orm';
+import { folders, snippet_tags, snippets } from '$lib/sqlite/schema'
+import { sql } from 'drizzle-orm'
 
 export type Snippet = {
   id: string
@@ -135,105 +120,6 @@ export type SnippetStore =
     remove(id: string): Promise<void>
     move(movingSnippet: Snippet, sourceFolderId: string, destinationFolderId: string): Promise<void>
   }
-
-export async function createLocalSnippetStoreV2(
-  migrationStateStore: Writable<MigrationState>,
-  globalStateStore: Writable<GlobalState>,
-  db: SqliteRemoteDatabase,
-): Promise<SnippetStore> {
-  let snippets: Snippet[] = []
-  let folderId = 'default'
-  const store = writable(snippets)
-  const stores = derived(
-    [globalStateStore, migrationStateStore],
-    ([globalState, migrationState]: [GlobalState, MigrationState]) => {
-      folderId = globalState.folderId
-      return [globalState, migrationState]
-    }
-  )
-  const pairUnsubscribeFn = stores.subscribe(
-    // TODO: fix the typing of `globalState` and `migrationState`, which both
-    //       have the type `GlobalState | MigrationState`
-    async ([globalState, migrationState]) => {
-      if (migrationState === 'done') {
-        const dbSnippets = await querySnippetsByFolderId(db, (globalState as GlobalState).folderId)
-        const snippetIds = dbSnippets.map(snippet => snippet.id)
-        const tags = await queryTagsBySnippetIds(db, snippetIds)
-        const tagsMap = buildTagsMap(tags)
-        snippets = dbSnippets.map(
-          (dbSnippet) => dbSnippetToDisplaySnippet(dbSnippet, tagsMap)
-        )
-        store.set(snippets)
-      }
-    }
-  )
-
-  return {
-    subscribe(
-      run: Subscriber<Snippet[]>,
-      invalidate?: Invalidator<Snippet[]>,
-    ): Unsubscriber {
-      const baseUnsubscribeFn = store.subscribe(run, invalidate)
-
-      return () => {
-        pairUnsubscribeFn()
-        baseUnsubscribeFn()
-      }
-    },
-    clone: async (snippet: Snippet) => {
-      const clonedSnippet: Snippet = {
-        ...snippet,
-        id: crypto.randomUUID(),
-        position: snippets.length + 1,
-        tags: snippet.tags.slice(),
-        createdAt: new Date().getTime(),
-        updatedAt: new Date().getTime(),
-      }
-      const dbSnippet = displaySnippetToDbSnippet(folderId, clonedSnippet)
-      await upsertSnippet(db, dbSnippet)
-      if (snippet.tags.length > 0) {
-        await upsertTags(db, clonedSnippet.id, snippet.tags)
-      }
-      snippets.push(clonedSnippet)
-
-      store.set(snippets)
-    },
-    upsert: async (snippet: Snippet) => {
-      const dbSnippet = displaySnippetToDbSnippet(folderId, snippet)
-      await upsertSnippet(db, dbSnippet)
-      if (snippet.tags && snippet.tags.length > 0) {
-        await clearTags(db, snippet.id)
-        await upsertTags(db, snippet.id, snippet.tags)
-      }
-      const index = snippets.findIndex(snippet_ => snippet_.id === snippet.id)
-      if (index === -1) {
-        snippets.push(snippet)
-        store.set(snippets)
-        return
-      }
-
-      snippets[index] = snippet
-      store.set(snippets)
-    },
-    remove: async (id: string) => {
-      await deleteSnippet_(db, id)
-      const index = snippets.findIndex(snippet => snippet.id === id)
-      snippets.splice(index, 1)
-
-      store.set(snippets)
-    },
-    move: async (movingSnippet: Snippet, sourceFolderId: string, destinationFolderId: string) => {
-      const dbSnippet = displaySnippetToDbSnippet(sourceFolderId, movingSnippet)
-      dbSnippet.folderId = destinationFolderId
-      await upsertSnippet(db, dbSnippet)
-
-      const index = snippets.findIndex(snippet => snippet.id === movingSnippet.id)
-      snippets.splice(index, 1)
-
-      store.set(snippets)
-    },
-  }
-}
 
 export async function readCatalog(): Promise<Catalog> {
   const opfsRoot = await navigator.storage.getDirectory()
